requirements.txt
awswrangler==3.6.0
boto3==1.34.49
botocore==1.34.49
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
cmdstanpy==1.2.1
contourpy==1.1.1
cycler==0.12.1
fonttools==4.49.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
gspread==6.0.2
gspread-dataframe==3.3.1
holidays==0.43
idna==3.6
importlib-resources==6.1.2
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.7.5
numpy==1.24.4
oauthlib==3.2.2
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pip==20.0.2
pkg-resources==0.0.0
prophet==1.1.5
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
setuptools==44.0.0
six==1.16.0
stanio==0.3.0
StrEnum==0.4.15
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
zipp==3.17.0
